+++ src/router/index.js
@@ -1,10 +1,11 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import Login from '../components/Login.vue'
-import Home from '../components/Home.vue'
-import welcome from '../components/welcome.vue'
-import AdminList from '../components/User/AdminList.vue'
-import MemberList from '../components/User/MemberList.vue'
+import Login from '../views/Login.vue'
+import Home from '../views/Home.vue'
+import welcome from '../views/welcome.vue'
+import AdminList from '../views/User/AdminList.vue'
+import MemberList from '../views/User/MemberList.vue'
+import securityShow from '../views/security-threejs/securityShow.vue'
 Vue.use(VueRouter)
 
 const router = new VueRouter({
@@ -24,7 +25,8 @@ const router = new VueRouter({
             children: [
                 { path: '/welcome', component: welcome },
                 { path: '/adminList', component: AdminList },
-                { path: '/memberList', component: MemberList }
+                { path: '/memberList', component: MemberList },
+                { path: '/BIMmodel', component: securityShow }
             ]
         }
     ]
